Revert "SM-5918"
This reverts commit 57ac9e0063e8409a9f9003a9d3048bc67aaf4ff6.

(cherry picked from commit 80657d4f27198e623c6ad541c0667551cc7b0628)

+++ i21Database/dbo/Tables/tblSMStartingNumber.sql
@@ -2,7 +2,7 @@
     [intStartingNumberId]                INT            IDENTITY (1, 1) NOT NULL,
     [strTransactionType]   NVARCHAR (100) COLLATE Latin1_General_CI_AS NOT NULL,
     [strPrefix]            NVARCHAR (50)  COLLATE Latin1_General_CI_AS NOT NULL,
-    [intNumber]            BIGINT            NOT NULL,
+    [intNumber]            INT            NOT NULL,
 	[intDigits]            INT            NULL,
     [strModule]            NVARCHAR (50)  COLLATE Latin1_General_CI_AS NOT NULL,
     [ysnUseLocation]       BIT            NOT NULL DEFAULT 0,
